Refactored service structure: moved TaskService, TaskServiceImpl to service/task/ and updated TaskController
